增加了高频彩近期开奖；内蒙快三和值提示；修改了双色球大额投注失败问题

+++ Ruyicai_General/main/v_main/src/com/ruyicai/activity/buy/InsufficientBalanceActivity.java
@@ -59,7 +59,7 @@ public class InsufficientBalanceActivity extends Activity {
 
 		// 初始化金额的显示
 		amtTextView = (TextView) findViewById(R.id.insufficient_balance_amt);
-		int menoy = Integer.valueOf(amountString) / 100;
+		long menoy = Long.valueOf(amountString) / 100;
 		amtTextView.setText(menoy + "元");
 
 		// 初始化直接支付按钮
